feat: Improved theming
