+++ Cargo.toml
@@ -8,7 +8,6 @@ repository = "https://github.com/alexm-dev/runa"
 documentation = "https://docs.rs/runa-tui"
 readme = "README.md"
 authors = ["alexm-dev"]
-rust-version = "1.85"
 categories = ["filesystem", "command-line-utilities", "command-line-interface"]
 keywords = ["file-browser", "file-manager", "tui", "cli", "search"]
 
@@ -21,7 +20,7 @@ dirs = "6.0.0"
 ratatui = "0.30.0"
 crossterm = "0.29.0"
 serde = { version = "1.0", features = ["derive"] }
-toml = "0.9.10"
+toml = "0.9"
 crossbeam-channel = "0.5"
 unicode-width = "0.2.2"
 humansize = "2.1.3"
@@ -45,7 +44,7 @@ opt-level = 3
 lto = true
 codegen-units = 1
 panic = "abort"
-strip = "debuginfo"
+strip = true
 debug = false
 
 [profile.release-windows]
