+++ src/config/theme.rs
@@ -5,8 +5,10 @@
 //!
 //! Also holds the internal themes and the logic to apply user overrides on top of them.
 
+use crate::config::presets::*;
 use crate::ui::widgets::{DialogPosition, DialogSize};
 use crate::utils::parse_color;
+
 use once_cell::sync::Lazy;
 use ratatui::style::{Color, Style};
 use serde::Deserialize;
@@ -158,6 +160,23 @@ impl Theme {
         self.symlink.or(Theme::internal_defaults().symlink)
     }
 
+    // Pane-specific style getters
+
+    pub fn parent_selection_style(&self) -> Style {
+        self.parent.selection_style(&self.selection)
+    }
+
+    pub fn preview_selection_style(&self) -> Style {
+        self.preview.selection_style(&self.selection)
+    }
+
+    pub fn preview_item_style(&self) -> Style {
+        self.preview.entry_style(&self.entry)
+    }
+    pub fn parent_item_style(&self) -> Style {
+        self.parent.entry_style(&self.entry)
+    }
+
     // Accessor methods for various theme properties
 
     pub fn selection_icon(&self) -> &str {
@@ -196,6 +215,18 @@ impl Theme {
             Some("catppuccin-frappe") => Some(catppuccin_frappe()),
             Some("catppuccin-macchiato") => Some(catppuccin_mocha()),
             Some("catppuccin-latte") => Some(catppuccin_latte()),
+
+            Some("nord") => Some(nord()),
+
+            Some("two-dark") => Some(two_dark()),
+            Some("one-dark") => Some(one_dark()),
+
+            Some("solarized-dark") => Some(solarized_dark()),
+            Some("solarized-light") => Some(solarized_light()),
+
+            Some("dracula") => Some(dracula()),
+
+            Some("monokai") => Some(monokai()),
 
             Some("nightfox") => Some(nightfox()),
             Some("carbonfox") => Some(carbonfox()),
@@ -238,6 +269,8 @@ impl Theme {
     fn map_to_bat_theme(internal_theme: &str) -> &'static str {
         match internal_theme {
             "default" => "TwoDark",
+            "two-dark" => "TwoDark",
+            "one-dark" => "OneHalfDark",
             "gruvbox-dark" | "gruvbox-dark-hard" | "gruvbox" => "gruvbox-dark",
             "gruvbox-light" => "gruvbox-light",
             "tokyonight-night" | "tokyonight" | "tokyonight-storm" => "TwoDark",
@@ -246,6 +279,11 @@ impl Theme {
             "catppuccin-macchiato" => "Catppuccin Macchiato",
             "catppuccin-mocha" | "catppuccin" => "Catppuccin Mocha",
             "nightfox" | "carbonfox" | "rose-pine" | "everforest" => "TwoDark",
+            "monokai" => "Monokai Extended (default)",
+            "nord" => "Nord",
+            "solarized-dark" => "Solarized (dark)",
+            "solarized-light" => "Solarized (light)",
+            "dracula" => "Dracula",
             _ => "TwoDark",
         }
     }
@@ -335,6 +373,7 @@ impl ColorPair {
 #[derive(Deserialize, Debug, PartialEq, Clone, Copy, Default)]
 #[serde(default)]
 pub struct PaneTheme {
+    #[serde(flatten)]
     color: ColorPair,
     selection: Option<ColorPair>,
 }
@@ -357,11 +396,15 @@ impl PaneTheme {
         }
     }
 
-    /// Returns the selection style, falling back to the internal default theme's selection style.
-    /// This method uses the internal default theme as the fallback.
-    pub fn selection_style_or_theme(&self) -> Style {
-        let fallback = Theme::internal_defaults().selection;
-        self.selection_style(&fallback)
+    /// Returns the entry style, falling back to the provided fallback ColorPair.
+    /// If entry color is Reset, uses the fallback.
+    ///
+    /// # Arguments
+    /// * `fallback` - A `ColorPair` to use as fallback.
+    /// # Returns
+    /// * `Style` - A `Style` representing the entry style.
+    pub fn entry_style(&self, fallback: &ColorPair) -> Style {
+        self.color.style_or(fallback)
     }
 
     /// Returns the pane color style, falling back to the provided fallback ColorPair.
@@ -444,17 +487,14 @@ pub struct WidgetTheme {
 }
 
 impl WidgetTheme {
-    /// Returns the dialog position.
     pub fn position(&self) -> &Option<DialogPosition> {
         &self.position
     }
 
-    /// Returns the dialog size.
     pub fn size(&self) -> &Option<DialogSize> {
         &self.size
     }
 
-    /// Returns the confirm dialog size.
     pub fn confirm_size(&self) -> &Option<DialogSize> {
         &self.confirm_size
     }
@@ -625,7 +665,6 @@ pub fn make_theme(name: &str, palette: Palette, icon: &str) -> Theme {
             fg: muted,
             ..ColorPair::default()
         },
-
         status_line: ColorPair {
             fg: Color::Reset,
             bg: base_bg,
@@ -667,167 +706,4 @@ pub fn make_theme(name: &str, palette: Palette, icon: &str) -> Theme {
         },
         ..Theme::default()
     }
-}
-
-// Theme palettes
-
-const TOKYO_STORM: Palette = Palette {
-    base: (36, 40, 59),
-    surface: (41, 46, 66),
-    overlay: (86, 95, 137),
-    primary: (187, 154, 247),
-    secondary: (125, 207, 255),
-    directory: (122, 162, 247),
-};
-
-const TOKYO_NIGHT: Palette = Palette {
-    base: (26, 27, 38),
-    surface: (44, 51, 78),
-    overlay: (86, 95, 137),
-    primary: (187, 154, 247),
-    secondary: (125, 207, 255),
-    directory: (122, 162, 247),
-};
-
-const TOKYO_DAY: Palette = Palette {
-    base: (225, 226, 231),
-    surface: (196, 199, 209),
-    overlay: (168, 175, 199),
-    primary: (152, 94, 171),
-    secondary: (52, 90, 183),
-    directory: (52, 90, 183),
-};
-
-pub fn tokyonight_storm() -> Theme {
-    make_theme("tokyonight-storm", TOKYO_STORM, "┃")
-}
-pub fn tokyonight_night() -> Theme {
-    make_theme("tokyonight-night", TOKYO_NIGHT, "┃")
-}
-pub fn tokyonight_day() -> Theme {
-    make_theme("tokyonight-day", TOKYO_DAY, "┃")
-}
-
-const GRUV_DARK_HARD: Palette = Palette {
-    base: (29, 32, 33),
-    surface: (60, 56, 54),
-    overlay: (146, 131, 116),
-    primary: (211, 134, 155),
-    secondary: (142, 192, 124),
-    directory: (131, 165, 152),
-};
-
-const GRUV_DARK: Palette = Palette {
-    base: (40, 40, 40),
-    surface: (60, 56, 54),
-    overlay: (146, 131, 116),
-    primary: (211, 134, 155),
-    secondary: (142, 192, 124),
-    directory: (131, 165, 152),
-};
-
-const GRUV_LIGHT: Palette = Palette {
-    base: (251, 241, 199),
-    surface: (213, 196, 161),
-    overlay: (124, 111, 100),
-    primary: (143, 63, 113),
-    secondary: (66, 123, 88),
-    directory: (7, 102, 120),
-};
-
-pub fn gruvbox_dark_hard() -> Theme {
-    make_theme("gruvbox-dark-hard", GRUV_DARK_HARD, "*")
-}
-pub fn gruvbox_dark() -> Theme {
-    make_theme("gruvbox-dark", GRUV_DARK, "*")
-}
-pub fn gruvbox_light() -> Theme {
-    make_theme("gruvbox-light", GRUV_LIGHT, "*")
-}
-
-const MOCHA: Palette = Palette {
-    base: (30, 30, 46),
-    surface: (49, 50, 68),
-    overlay: (108, 112, 134),
-    primary: (203, 166, 247),
-    secondary: (148, 226, 213),
-    directory: (137, 180, 250),
-};
-
-const FRAPPE: Palette = Palette {
-    base: (48, 52, 70),
-    surface: (65, 69, 89),
-    overlay: (115, 121, 148),
-    primary: (202, 158, 230),
-    secondary: (129, 200, 190),
-    directory: (140, 170, 238),
-};
-
-const LATTE: Palette = Palette {
-    base: (239, 241, 245),
-    surface: (204, 208, 218),
-    overlay: (156, 160, 176),
-    primary: (136, 57, 239),
-    secondary: (23, 146, 153),
-    directory: (30, 102, 245),
-};
-
-pub fn catppuccin_mocha() -> Theme {
-    make_theme("catppuccin-mocha", MOCHA, "┃")
-}
-pub fn catppuccin_frappe() -> Theme {
-    make_theme("catppuccin-frappe", FRAPPE, "┃")
-}
-pub fn catppuccin_latte() -> Theme {
-    make_theme("catppuccin-latte", LATTE, "┃")
-}
-
-const CARBON: Palette = Palette {
-    base: (22, 22, 22),
-    surface: (42, 42, 42),
-    overlay: (82, 82, 82),
-    primary: (190, 149, 233),
-    secondary: (61, 187, 199),
-    directory: (120, 169, 235),
-};
-
-const NIGHTFOX: Palette = Palette {
-    base: (25, 30, 36),
-    surface: (43, 51, 63),
-    overlay: (87, 91, 112),
-    primary: (195, 157, 239),
-    secondary: (99, 199, 209),
-    directory: (113, 161, 236),
-};
-
-pub fn carbonfox() -> Theme {
-    make_theme("carbonfox", CARBON, "┃")
-}
-pub fn nightfox() -> Theme {
-    make_theme("nightfox", NIGHTFOX, "┃")
-}
-
-const FOREST: Palette = Palette {
-    base: (43, 51, 57),
-    surface: (74, 82, 88),
-    overlay: (133, 146, 137),
-    primary: (167, 192, 128),
-    secondary: (230, 126, 128),
-    directory: (127, 187, 179),
-};
-
-const ROSE_PINE: Palette = Palette {
-    base: (25, 23, 36),
-    surface: (31, 29, 46),
-    overlay: (110, 106, 134),
-    primary: (196, 167, 231),
-    secondary: (235, 188, 186),
-    directory: (49, 116, 143),
-};
-
-pub fn everforest() -> Theme {
-    make_theme("everforest", FOREST, "*")
-}
-pub fn rose_pine() -> Theme {
-    make_theme("rose_pine", ROSE_PINE, "*")
 }
